pyproject.toml
[tool.poetry]
name = "ruslan-schedule"
version = "0.1.0"
description = ""
authors = ["Anzor <magometov_a.z@mail.ru>"]
readme = "README.md"
packages = [{include = "ruslan_schedule"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
djangorestframework-camel-case = "^1.4.2"
python-dotenv = "^1.0.0"
sentry-sdk = {extras = ["django"], version = "^1.25.0"}


[tool.poetry.group.development.dependencies]
django-stubs = "^4.2.1"
djangorestframework-stubs = "^3.14.1"
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
